Crate forest_blocks

Structs

Block
A complete block

BlockHeader
Header of a block

BlockHeaderBuilder
Builder for BlockHeader.

EPostProof
Proof of Spacetime election proof

EPostTicket
PoSt election candidates

FullTipset
FullTipSet is an expanded version of the TipSet that contains all the blocks and messages

Ticket
A Ticket is a marker of a tick of the blockchain's clock. It is the source of randomness for proofs of storage and leader election. It is generated by the miner of a block using a VRF and a VDF.

TipSetKeys
A set of CIDs forming a unique key for a TipSet. Equal keys will have equivalent iteration order, but note that the CIDs are not maintained in the same order as the canonical iteration order of blocks in a tipset (which is by ticket)

Tipset
An immutable set of blocks at the same height with the same parent set. Blocks in a tipset are canonically ordered by ticket size.

TxMeta
Tracks the merkleroots of both secp and bls messages separately

Enums

Error
Blockchain blocks error