Module forest_blocks::ticket

Modules

json

Structs

EPostProof
Proof of Spacetime election proof

EPostTicket
PoSt election candidates

Ticket
A Ticket is a marker of a tick of the blockchain's clock. It is the source of randomness for proofs of storage and leader election. It is generated by the miner of a block using a VRF and a VDF.