Crate forest_blocks

Re-exports

pub use election_proof::*;

pub use gossip_block::*;

pub use header::*;

pub use ticket::*;

pub use tipset::*;

Modules

election_proof

gossip_block

header

ticket

tipset

Structs

Block
A complete Filecoin block. This contains the block header as well as all bls and secp messages.

TxMeta
Tracks the merkleroots of both secp and bls messages separately.

Enums

Error
Blockchain blocks error

Constants

BLOCK_MESSAGE_LIMIT
Limit of bls and secp messages combined in a block.