Struct forest_blocks::header::BlockHeader

pub struct BlockHeader {
    pub signature: Option<Signature>,
    /* private fields */
}

Header of a block

Usage:

use forest_blocks::{BlockHeader, TipsetKeys, Ticket};
use fvm_shared::address::Address;
use cid::Cid;
use cid::multihash::Code::Identity;
use fvm_shared::bigint::BigInt;
use fvm_shared::crypto::signature::Signature;
use fvm_ipld_encoding::DAG_CBOR;
use cid::multihash::MultihashDigest;

BlockHeader::builder()
    .messages(Cid::new_v1(DAG_CBOR, Identity.digest(&[]))) // required
    .message_receipts(Cid::new_v1(DAG_CBOR, Identity.digest(&[]))) // required
    .state_root(Cid::new_v1(DAG_CBOR, Identity.digest(&[]))) // required
    .miner_address(Address::new_id(0)) // optional
    .beacon_entries(Vec::new()) // optional
    .winning_post_proof(Vec::new()) // optional
    .election_proof(None) // optional
    .bls_aggregate(None) // optional
    .signature(None) // optional
    .parents(TipsetKeys::default()) // optional
    .weight(BigInt::from(0u8)) // optional
    .epoch(0) // optional
    .timestamp(0) // optional
    .ticket(Some(Ticket::default())) // optional
    .fork_signal(0) // optional
    .build()
    .unwrap();

Fields
signature: Option<Signature>
Implementations

impl BlockHeader

pub fn builder() -> BlockHeaderBuilder

Generates a BlockHeader builder as a constructor

pub fn parents(&self) -> &TipsetKeys

Get BlockHeader parents

pub fn weight(&self) -> &BigInt

Get BlockHeader weight

pub fn epoch(&self) -> ChainEpoch

Get BlockHeader epoch

pub fn beacon_entries(&self) -> &[BeaconEntry]

Get Drand BeaconEntry

pub fn winning_post_proof(&self) -> &[PoStProof]

Get winning PoSt proof

pub fn miner_address(&self) -> &Address

Get BlockHeader.miner_address

pub fn messages(&self) -> &Cid

Get BlockHeader.messages

pub fn message_receipts(&self) -> &Cid

Get BlockHeader.message_receipts

pub fn state_root(&self) -> &Cid

Get BlockHeader.state_root

pub fn timestamp(&self) -> u64

Get BlockHeader.timestamp

pub fn ticket(&self) -> &Option<Ticket>

Get BlockHeader.ticket

pub fn bls_aggregate(&self) -> &Option<Signature>

Get BlockHeader.bls_aggregate

pub fn cid(&self) -> &Cid

Get BlockHeader.cid

pub fn parent_base_fee(&self) -> &BigInt

Get BlockHeader.parent_base_fee

pub fn fork_signal(&self) -> u64

Get BlockHeader.fork_signal

pub fn election_proof(&self) -> &Option<ElectionProof>

Get BlockHeader.election_proof

pub fn signature(&self) -> &Option<Signature>

Get BlockHeader.signature

pub fn to_sort_key(&self) -> Option<([u8; 32], Vec<u8>)>

Key used for sorting headers and blocks.

pub fn check_block_signature(&self, addr: &Address) -> Result<(), Error>

Check to ensure block signature is valid

pub fn is_ticket_winner(ticket: &Ticket, mpow: BigInt, net_pow: BigInt) -> bool

Returns true if (h(vrfout) * totalPower) < (e * sectorSize * 2^256)

pub async fn validate_block_drand<B: Beacon>(
&self,
network_version: NetworkVersion,
b_schedule: &BeaconSchedule<B>,
parent_epoch: ChainEpoch,
prev_entry: &BeaconEntry
) -> Result<(), Error>

Validates if the current header’s Beacon entries are valid to ensure randomness was generated correctly

pub fn to_signing_bytes(&self) -> Vec<u8>

Serializes the header to bytes for signing purposes i.e. without the signature field