Struct forest_blocks::header::BlockHeaderBuilder

pub struct BlockHeaderBuilder { /* fields omitted */ }

Builder for BlockHeader.

Implementations

impl BlockHeaderBuilder

pub fn parents(&mut self, value: TipsetKeys) -> &mut Self

Parents is the set of parents this block was based on. Typically one, but can be several in the case where there were multiple winning ticket- holders for an epoch

pub fn weight(&mut self, value: BigInt) -> &mut Self

weight is the aggregate chain weight of the parent set

pub fn epoch(&mut self, value: ChainEpoch) -> &mut Self

epoch is the period in which a new block is generated. There may be multiple rounds in an epoch.

pub fn beacon_entries(&mut self, value: Vec<BeaconEntry>) -> &mut Self

BeaconEntries contain the verifiable oracle randomness used to elect this block’s author leader

pub fn winning_post_proof(&mut self, value: Vec<PoStProof>) -> &mut Self

PoStProofs are the winning post proofs

pub fn miner_address(&mut self, value: Address) -> &mut Self

miner_address is the address of the miner actor that mined this block

pub fn messages(&mut self, value: Cid) -> &mut Self

messages contains the Cid to the merkle links for bls_messages and secp_messages

pub fn message_receipts(&mut self, value: Cid) -> &mut Self

message_receipts is the Cid of the root of an array of MessageReceipts

pub fn state_root(&mut self, value: Cid) -> &mut Self

state_root is a cid pointer to the parent state root after calculating parent tipset.

pub fn fork_signal(&mut self, value: u64) -> &mut Self

pub fn signature(&mut self, value: Option<Signature>) -> &mut Self

pub fn election_proof(&mut self, value: Option<ElectionProof>) -> &mut Self

pub fn timestamp(&mut self, value: u64) -> &mut Self

timestamp, in seconds since the Unix epoch, at which this block was created

pub fn ticket(&mut self, value: Option<Ticket>) -> &mut Self

the ticket submitted with this block

pub fn bls_aggregate(&mut self, value: Option<Signature>) -> &mut Self

aggregate signature of miner in block

pub fn parent_base_fee(&mut self, value: TokenAmount) -> &mut Self

the base fee of the parent block

pub fn build(&self) -> Result<BlockHeader, String>

Builds a new BlockHeader.

Errors

If a required field has not been initialized.